admin/questions/controllerの生成
