+++ app/controllers/application_controller.rb
@@ -4,6 +4,6 @@ class ApplicationController < ActionController::Base
   protected
 
   def configure_permitted_parameters
-    devise_parameter_sanitizer.permit(:sign_up, keys: [:name, :age])
+    devise_parameter_sanitizer.permit(:sign_up, keys: [:name, :birthday])
   end
 end
