+++ app/helpers/admin/questions_helper.rb
@@ -1,0 +1,2 @@
+module Admin::QuestionsHelper
+end
