+++ app/views/devise/registrations/new.html.haml
@@ -5,9 +5,9 @@
     = f.label :name
     %br/
     = f.text_field :name
-  .field
-    = f.label :age
-    %br/
+  -# .field
+  -#   = f.label :age
+  -#   %br/
     = f.number_field :age
   .field
     = f.label :email
