+++ app/views/home/index.html.haml
@@ -2,6 +2,7 @@
 - if user_signed_in?
   %h2 ログインしています
   = link_to "ログアウト", destroy_user_session_path, method: :delete
+  = link_to "質問登録", new_admin_question_path
 - else
   %h2 ログインしていません
   = link_to "新規登録", new_user_registration_path
