+++ config/routes.rb
@@ -1,6 +1,10 @@
 Rails.application.routes.draw do
   devise_for :users
   root to: "home#index"
-  
-  resources :questions, except: :show
+
+  resources :restaurants, only: :index 
+  namespace :admin do
+    resources :questions, except: :show
+  end
 end
+
