+++ test/controllers/admin/questions_controller_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class Admin::QuestionsControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
